Move podNIC code to podnic.go
Currectly podinterface.go has a mix of code related to podnic struct and
the bindings. This change move the code realted to podNIC to podnic.go
